Further improve view models
