+++ AirCasting/SearchAndFollow/CompleteScreen/CompleteScreenViewModel.swift
@@ -4,6 +4,12 @@
 import Foundation
 import SwiftUI
 
+struct SessionStreamViewModel: Identifiable {
+    let id: Int
+    let sensorName: String
+    let lastMeasurementValue: Double
+}
+
 class CompleteScreenViewModel: ObservableObject {
     private let session: SearchSession
     @Published var selectedStream: Int?
@@ -14,7 +20,7 @@ class CompleteScreenViewModel: ObservableObject {
     let sessionStartTime: Date
     let sessionEndTime: Date
     let sensorType: String
-    let sessionStreams: [SearchSession.SearchSessionStream]
+    let sessionStreams: [SessionStreamViewModel]
     @Published var streamForChart: SearchSession.SearchSessionStream?
     
     init(session: SearchSession) {
@@ -24,7 +30,7 @@ class CompleteScreenViewModel: ObservableObject {
         sessionName = session.name
         sessionStartTime = session.startTime
         sessionEndTime = session.endTime
-        sessionStreams = session.streams
+        sessionStreams = session.streams.map({ .init(id: $0.id, sensorName: Self.showStreamName($0.sensorName), lastMeasurementValue: $0.measurements.last?.value ?? 0) })
         selectedStream = session.streams.first?.id
         streamForChart = session.streams.first
         sensorType = "OpenAir"
@@ -36,5 +42,16 @@ class CompleteScreenViewModel: ObservableObject {
     
     func chartTapped() {
         isMapSelected.toggle()
+    }
+    
+    func selectedStream(with id: Int) {
+        selectedStream = id
+    }
+    
+    private static func showStreamName(_ streamName: String) -> String {
+        streamName
+            .replacingOccurrences(of: ":", with: "-")
+            .drop { $0 != "-" }
+            .replacingOccurrences(of: "-", with: "")
     }
 }
