+++ AirCasting/SearchAndFollow/CompleteScreen/StaticSingleStreamView.swift
@@ -4,26 +4,23 @@
 import SwiftUI
 
 struct StaticSingleStreamView: View {
-    @StateObject var viewModel: StaticSingleStreamViewModel
-    @Binding var selectedStreamId: Int?
-    
-    init(selectedStreamId: Binding<Int?>,streamId: Int, streamName: String, value: Double) {
-        _selectedStreamId = .init(projectedValue: selectedStreamId)
-        _viewModel = .init(wrappedValue: StaticSingleStreamViewModel(streamId: streamId, streamName: streamName, value: value))
-    }
+    let streamName: String
+    let value: Double
+    let isSelected: Bool
+    let action: () -> Void
     
     var body: some View {
         VStack(spacing: 3) {
             Button(action: {
-                selectedStreamId = viewModel.streamId
+                action()
             }, label: {
                 VStack(spacing: 1) {
-                    Text(viewModel.streamName)
+                    Text(streamName)
                         .font(Fonts.systemFont1)
                         .scaledToFill()
                         HStack(spacing: 3) {
                             dot
-                            Text("\(Int(viewModel.value))")
+                            Text("\(Int(value))")
                                 .font(Fonts.regularHeading3)
                                 .scaledToFill()
                         }
@@ -31,7 +28,7 @@ struct StaticSingleStreamView: View {
                         .padding(.horizontal, 9)
                         .overlay(
                             RoundedRectangle(cornerRadius: 8)
-                                .strokeBorder((selectedStreamId == viewModel.streamId) ? Color.aircastingGray : .clear)
+                                .strokeBorder(isSelected ? Color.aircastingGray : .clear)
                         )
                     }
             })
@@ -47,6 +44,6 @@ struct StaticSingleStreamView: View {
 
 struct StaticSingleStreamView_Previews: PreviewProvider {
     static var previews: some View {
-        StaticSingleStreamView(selectedStreamId: .constant(3), streamId: 3, streamName: "AirBeam3-PM1", value: 20)
+        StaticSingleStreamView(streamName: "AirBeam3-PM1", value: 20, isSelected: true) { }
     }
 }
